move the components
