+++ components/getCompany.tsx
