+++ tsconfig.json
@@ -22,7 +22,12 @@
       {
         "name": "next"
       }
-    ]
+    ],"baseUrl": ".",
+    "paths": {
+      "@/components/*":["components/*"],
+      "@global-style":["styles/globals.css"],
+      "@/public/*":["public/*"]
+    }
   },
   "include": [
     "next-env.d.ts",
